libamxb: v4.11.8: Always log bus timeouts to syslog
- ddd0af980d88 ("[CI] Update changelog for release master_v4.11.8")
- ea8a01b78bac ("Issue: HOP-8243 [Amxb] Always log bus timeouts to syslog")

A logging function can be set in lib amxb, if this is done all the amxb calls involving bus access will be logged using this function.
In order to always log bus timeouts , use syslog instead of the custom logging function when a timeout has occured. This increases visibility of bus timeouts.

Signed-off-by: Hela Saadani <hela.saadani_ext@softathome.com>